% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_bar.R
\name{facet_bar}
\alias{facet_bar}
\title{facet_bar}
\usage{
facet_bar(df, y, x, by, nrow = 2, ncol = 3, scales = "free")
}
\arguments{
\item{df}{data frame used for plotting}

\item{y}{what to plot on the y-axis}

\item{x}{what to plot on the x-axis}

\item{by}{the grouping/faceting variable}

\item{nrow}{number of rows for the plot}

\item{ncol}{number of panels in each row}

\item{scales}{"free" allows for each panel to have its own scales}
}
\value{

}
\description{
facet_bar
}
