% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_visual.R
\name{LDA_visual}
\alias{LDA_visual}
\title{LDA_visual
This function helps transform textmineR output to be usable for the LDAvis package}
\usage{
LDA_visual(model, data_tokenised)
}
\arguments{
\item{model}{the model output file from the textmineR package}

\item{data_tokenised}{a data frame that contains three cols: document, word, n}
}
\value{

}
\description{
LDA_visual
This function helps transform textmineR output to be usable for the LDAvis package
}
\examples{

}
