% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm101-package.R
\docType{package}
\name{tm101-package}
\alias{tm101}
\alias{tm101-package}
\title{tm101: What the Package Does (One Line, Title Case)}
\description{
This is the package dedicated to my intro to text mining modules at Cambridge Digital Humanities.
}
\author{
\strong{Maintainer}: Meng Liu \email{veronica.lm625@gmail.com} (\href{https://orcid.org/0000-0001-8323-2699}{ORCID})

}
\keyword{internal}
